+++ build-resources/geocml-task-scheduler/geocml-task-scheduler/schedule.py
@@ -1,7 +1,7 @@
 from task import Task
-from tasks.backup_geocml_db import backup_geocml_db
-from tasks.restore_geocml_db_from_backups import restore_geocml_db_from_backups
-from tasks.healthcheck_services import healthcheck_services
+from backup_geocml_db import backup_geocml_db
+from restore_geocml_db_from_backups import restore_geocml_db_from_backups
+from healthcheck_services import healthcheck_services
 
 backup_geocml_db_task = Task(backup_geocml_db, (), 3600) # task runs every hour
 backup_geocml_db_task.start()
