+++ build-resources/geocml-task-scheduler/geocml-task-scheduler/task.py
@@ -22,15 +22,15 @@ class Task(Thread):
             sleep(self.timeout)
             start = time()
             log('Running \'{}\' at {}'.format(self.target.__name__, start), self.target)
-        
-            try: 
+
+            try:
                 ret_code = self.target()
                 log('Ran \'{}\' in {}'.format(self.target.__name__, time() - start), self.target)
                 if ret_code == 0: # ret_code 0: task stopped successfully
                     log('Stopping task \'{}\' at {} with exit code 0'.format(self.target.__name__, time()),
                        self.target)
                     return
-            # TODO: ret_code 1: task stopped due to expected error  
+            # TODO: ret_code 1: task stopped due to expected error
             # TODO: other ret_codes: log warning (ret_code {} is not valid)
             except Exception as e:
                 log('FATAL: {} at {}'.format(e, time())) # TODO: should the task be stopped after this?
